+++ dependencyInjection/src/main/java/com/example/dependencyInjection/model/MidDeveloper.java
@@ -1,7 +1,7 @@
 package com.example.dependencyInjection.model;
 
 public class MidDeveloper extends Developer{
-    public MidDeveloper(int id, String name, int salary) {
+    public MidDeveloper(int id, String name, double salary) {
         super(id, name, salary, Experience.MID);
     }
 }
