+++ dependencyInjection/src/main/java/com/example/dependencyInjection/model/SeniorDeveloper.java
@@ -1,7 +1,7 @@
 package com.example.dependencyInjection.model;
 
 public class SeniorDeveloper extends Developer {
-    public SeniorDeveloper(int id, String name, int salary) {
+    public SeniorDeveloper(int id, String name, double salary) {
         super(id, name, salary, Experience.SENIOR);
     }
 }
